LayoutTests/fast/scrolling/ios/scrollbar-hiding-expected.txt

Both


Just horizontal


Just vertical

None visible


(scrolling tree
  (frame scrolling node
    (scrollable area size width=800 height=600)
    (total content size width=800 height=600)
    (last committed scroll position (0,0))
    (scrollable area parameters
      (horizontal scroll elasticity 1)
      (vertical scroll elasticity 1)
      (horizontal scrollbar mode 0)
      (vertical scrollbar mode 0)
      (horizontal scrollbar hidden by style 1)
      (vertical scrollbar hidden by style 1))
    (layout viewport (0,0) width=800 height=600)
    (min layoutViewport origin (0,0))
    (max layoutViewport origin (0,0))
    (behavior for fixed 0)
    (overflow scrolling node
      (scrollable area size width=200 height=100)
      (total content size width=400 height=200)
      (last committed scroll position (0,0))
      (scrollable area parameters
        (horizontal scroll elasticity 1)
        (vertical scroll elasticity 1)
        (horizontal scrollbar mode 0)
        (vertical scrollbar mode 0)
        (has enabled horizontal scrollbar 1)
        (has enabled vertical scrollbar 1)))
    (overflow scrolling node
      (scrollable area size width=200 height=100)
      (total content size width=400 height=200)
      (last committed scroll position (0,0))
      (scrollable area parameters
        (horizontal scroll elasticity 1)
        (vertical scroll elasticity 1)
        (horizontal scrollbar mode 0)
        (vertical scrollbar mode 0)
        (has enabled horizontal scrollbar 1)
        (has enabled vertical scrollbar 1)
        (vertical scrollbar hidden by style 1)))
    (overflow scrolling node
      (scrollable area size width=200 height=100)
      (total content size width=400 height=200)
      (last committed scroll position (0,0))
      (scrollable area parameters
        (horizontal scroll elasticity 1)
        (vertical scroll elasticity 1)
        (horizontal scrollbar mode 0)
        (vertical scrollbar mode 0)
        (has enabled horizontal scrollbar 1)
        (has enabled vertical scrollbar 1)
        (horizontal scrollbar hidden by style 1)))
    (overflow scrolling node
      (scrollable area size width=200 height=100)
      (total content size width=400 height=200)
      (last committed scroll position (0,0))
      (scrollable area parameters
        (horizontal scroll elasticity 1)
        (vertical scroll elasticity 1)
        (horizontal scrollbar mode 0)
        (vertical scrollbar mode 0)
        (has enabled horizontal scrollbar 1)
        (has enabled vertical scrollbar 1)
        (horizontal scrollbar hidden by style 1)
        (vertical scrollbar hidden by style 1)))))
